Moved class
- Moved the Header class,
