+++ app/Http/Controllers/ResourceView.php
@@ -8,7 +8,7 @@ use App\Models\ResourceType;
 use App\Option\ResourceCollection;
 use App\Option\ResourceItem;
 use App\Request\Parameter;
-use App\Response\Header\Header;
+use App\Response\Header;
 use App\Response\Pagination as UtilityPagination;
 use App\Transformers\Resource as ResourceTransformer;
 use Illuminate\Http\JsonResponse;
